orc/pyproject.toml

[tool.poetry]
name = "orc"
version = "1.0.0"
description = "Codebase Intelligence Agent"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.0"
rich = "^10.0.0"
networkx = "^2.6.0"
pyyaml = "^5.4.0"
flask = "^2.0.0"
dataclasses-json = "^0.5"

[tool.poetry.scripts]
orc = "run_orc:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"